update batch implemented

+++ src/math/Sigmoid.hpp
@@ -5,11 +5,11 @@
 #include <cmath>
 class Sigmoid : ActivationFunction {
 private:
-  static float sigmoid(float x) { return 1.0f / (1 + std::expf(-x)); }
+  static float sigmoid(float x) { return 1.0f / (1 + std::exp(-x)); }
   static float sigmoid_deriv(float x) {
     //(1 + e^-x)^-1
     //-(1+e^(-x))^(-2) * e^(-x) * (-1)
-    return std::pow(1 + std::expf(-x), -2) * std::expf(-x);
+    return std::pow(1 + std::exp(-x), -2) * std::exp(-x);
   }
 
 public:
